pyproject.toml
[project]
name = "callback-api"
version = "0.1.0"
description = ""
authors = [
    {name = "lil-marsxlll",email = "d.marsele18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "fastapi[all] (>=0.135.1,<0.136.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "black (>=26.3.0,<27.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
